+++ src/Symfony/Component/Console/Descriptor/JsonDescriptor.php
@@ -33,6 +33,9 @@ class JsonDescriptor extends Descriptor
 
     protected function describeInputOption(InputOption $option, array $options = []): void
     {
+        if ($this->skipHiddenOption($option, $options)) {
+            return;
+        }
         $this->writeData($this->getInputOptionData($option), $options);
         if ($option->isNegatable()) {
             $this->writeData($this->getInputOptionData($option, true), $options);
@@ -41,12 +44,12 @@ class JsonDescriptor extends Descriptor
 
     protected function describeInputDefinition(InputDefinition $definition, array $options = []): void
     {
-        $this->writeData($this->getInputDefinitionData($definition), $options);
+        $this->writeData($this->getInputDefinitionData($definition, $options), $options);
     }
 
     protected function describeCommand(Command $command, array $options = []): void
     {
-        $this->writeData($this->getCommandData($command, $options['short'] ?? false), $options);
+        $this->writeData($this->getCommandData($command, $options), $options);
     }
 
     protected function describeApplication(Application $application, array $options = []): void
@@ -56,7 +59,7 @@ class JsonDescriptor extends Descriptor
         $commands = [];
 
         foreach ($description->getCommands() as $command) {
-            $commands[] = $this->getCommandData($command, $options['short'] ?? false);
+            $commands[] = $this->getCommandData($command, $options);
         }
 
         $data = [];
@@ -101,26 +104,32 @@ class JsonDescriptor extends Descriptor
 
     private function getInputOptionData(InputOption $option, bool $negated = false): array
     {
-        return $negated ? [
-            'name' => '--no-'.$option->getName(),
+        $data = $negated ? [
+            'name' => '--no-' . $option->getName(),
             'shortcut' => '',
             'accept_value' => false,
             'is_value_required' => false,
             'is_multiple' => false,
-            'description' => 'Negate the "--'.$option->getName().'" option',
+            'is_deprecated' => $option->isDeprecated(),
+            'description' => 'Negate the "--' . $option->getName() . '" option',
             'default' => false,
         ] : [
-            'name' => '--'.$option->getName(),
-            'shortcut' => $option->getShortcut() ? '-'.str_replace('|', '|-', $option->getShortcut()) : '',
+            'name' => '--' . $option->getName(),
+            'shortcut' => $option->getShortcut() ? '-' . str_replace('|', '|-', $option->getShortcut()) : '',
             'accept_value' => $option->acceptValue(),
             'is_value_required' => $option->isValueRequired(),
             'is_multiple' => $option->isArray(),
+            'is_deprecated' => $option->isDeprecated(),
             'description' => preg_replace('/\s*[\r\n]\s*/', ' ', $option->getDescription()),
             'default' => \INF === $option->getDefault() ? 'INF' : $option->getDefault(),
         ];
+        if (!$option->isDeprecated()) {
+            unset($data['is_deprecated']);
+        }
+        return $data;
     }
 
-    private function getInputDefinitionData(InputDefinition $definition): array
+    private function getInputDefinitionData(InputDefinition $definition, array $options): array
     {
         $inputArguments = [];
         foreach ($definition->getArguments() as $name => $argument) {
@@ -128,7 +137,7 @@ class JsonDescriptor extends Descriptor
         }
 
         $inputOptions = [];
-        foreach ($definition->getOptions() as $name => $option) {
+        foreach ($this->removeHiddenOptions($definition->getOptions(), $options) as $name => $option) {
             $inputOptions[$name] = $this->getInputOptionData($option);
             if ($option->isNegatable()) {
                 $inputOptions['no-'.$name] = $this->getInputOptionData($option, true);
@@ -138,8 +147,10 @@ class JsonDescriptor extends Descriptor
         return ['arguments' => $inputArguments, 'options' => $inputOptions];
     }
 
-    private function getCommandData(Command $command, bool $short = false): array
+    private function getCommandData(Command $command, array $options = []): array
     {
+        $short = $options['short'] ?? false;
+
         $data = [
             'name' => $command->getName(),
             'description' => $command->getDescription(),
@@ -155,7 +166,7 @@ class JsonDescriptor extends Descriptor
             $data += [
                 'usage' => array_merge([$command->getSynopsis()], $command->getUsages(), $command->getAliases()),
                 'help' => $command->getProcessedHelp(),
-                'definition' => $this->getInputDefinitionData($command->getDefinition()),
+                'definition' => $this->getInputDefinitionData($command->getDefinition(), $options),
             ];
         }