+++ src/Symfony/Component/Console/CHANGELOG.md
@@ -11,6 +11,8 @@ CHANGELOG
 ---
 
  * Add `ArgvInput::getRawTokens()`
+ * Add `InputOption::HIDDEN` flag to hide options
+ * Add `InputOption::DEPRECATED` flag to mark options as deprecated
 
 7.0
 ---